+++ jambeth/jambeth-cache/src/main/java/com/koch/ambeth/cache/collections/CacheMapEntryTypeProvider.java
@@ -42,16 +42,17 @@ public class CacheMapEntryTypeProvider implements ICacheMapEntryTypeProvider {
 	@Autowired
 	protected IAccessorTypeProvider accessorTypeProvider;
 
-	protected final Tuple2KeyHashMap<Class<?>, Byte, ICacheMapEntryFactory> typeToConstructorMap = new Tuple2KeyHashMap<>();
+	protected final Tuple2KeyHashMap<Class<?>, Integer, ICacheMapEntryFactory> typeToConstructorMap =
+			new Tuple2KeyHashMap<>();
 
 	protected final Lock writeLock = new ReentrantLock();
 
 	@Override
-	public ICacheMapEntryFactory getCacheMapEntryType(Class<?> entityType, byte idIndex) {
+	public ICacheMapEntryFactory getCacheMapEntryType(Class<?> entityType, int idIndex) {
 		if (bytecodeEnhancer == null) {
 			return ci;
 		}
-		ICacheMapEntryFactory factory = typeToConstructorMap.get(entityType, Byte.valueOf(idIndex));
+		ICacheMapEntryFactory factory = typeToConstructorMap.get(entityType, idIndex);
 		if (factory != null) {
 			return factory;
 		}
@@ -59,7 +60,7 @@ public class CacheMapEntryTypeProvider implements ICacheMapEntryTypeProvider {
 		writeLock.lock();
 		try {
 			// concurrent thread might have been faster
-			factory = typeToConstructorMap.get(entityType, Byte.valueOf(idIndex));
+			factory = typeToConstructorMap.get(entityType, idIndex);
 			if (factory != null) {
 				return factory;
 			}
@@ -88,7 +89,7 @@ public class CacheMapEntryTypeProvider implements ICacheMapEntryTypeProvider {
 			finally {
 				currentThread.setContextClassLoader(oldClassLoader);
 			}
-			typeToConstructorMap.put(entityType, Byte.valueOf(idIndex), factory);
+			typeToConstructorMap.put(entityType, idIndex, factory);
 			return factory;
 		}
 		finally {
